Extract Section component from App
